Allow fractional tied ranks in research model

+++ src/ncaa_rankings/models/modern_legacy.py
@@ -35,7 +35,7 @@ class ModernLegacyModel:
     def score_game(
         self,
         game: TeamGame,
-        opponent_rank: int | None,
+        opponent_rank: float | None,
         team_count: int,
         *,
         age_weeks: float = 0.0,
